lib smb

+++ lib/smb.c
@@ -505,7 +505,7 @@ static CURLcode smb_recv_message(struct Curl_easy *data, void **msg)
   if(!bytes_read)
     return CURLE_OK;
 
-  smbc->got += bytes_read;
+  smbc->got += (size_t)bytes_read;
 
   /* Check for a 32-bit nbt header */
   if(smbc->got < sizeof(unsigned int))
@@ -671,7 +671,7 @@ static CURLcode smb_send_setup(struct Curl_easy *data)
   MSGCATNULL(smbc->domain);
   MSGCATNULL(OS);
   MSGCATNULL(CLIENTNAME);
-  byte_count = p - msg.bytes;
+  byte_count = (size_t)(p - msg.bytes);
   msg.byte_count = smb_swap16((unsigned short)byte_count);
 
   return smb_send_message(data, SMB_COM_SETUP_ANDX, &msg,
@@ -699,7 +699,7 @@ static CURLcode smb_send_tree_connect(struct Curl_easy *data)
   MSGCAT("\\");
   MSGCATNULL(smbc->share);
   MSGCATNULL(SERVICENAME); /* Match any type of service */
-  byte_count = p - msg.bytes;
+  byte_count = (size_t)(p - msg.bytes);
   msg.byte_count = smb_swap16((unsigned short)byte_count);
 
   return smb_send_message(data, SMB_COM_TREE_CONNECT_ANDX, &msg,
